+++ ks44team04/src/main/resources/templates/fragments/leftmenubum.html
@@ -4,8 +4,10 @@
 
 <li th:fragment="leftmenubum"><a><i class="fa fa-edit"></i> 주문관리 <span class="fa fa-chevron-down"></span></a>
     <ul class="nav child_menu">
-        <li><a th:href="@{/admin/addresslist}">배송지목록</a></li>
-        <li><a th:href="@{/admin/postlist}">배송정보목록</a></li>
+        <li><a th:href="@{/user/address/list}">배송지목록</a></li>
+        <li><a th:href="@{/user/address/register}">배송지등록</a></li>
+        <li><a th:href="@{/user/address/modify}">배송지수정</a></li>
+        <li><a th:href="@{/user/address/delete}">배송지삭제</a></li>
     </ul>
 </li>
